+++ src/nodes/generate_final_answer.py
@@ -2,7 +2,7 @@ from configs.prompts import ANSWER_GENERATION_TEMPLATE
 
 from .base_node import BaseNode
 
-from ..models.graph_state import GraphState
+from ..modules.graph_state import GraphState
 
 
 class GenerateAnswerNode(BaseNode):
