+++ src/nodes/get_example.py
@@ -1,6 +1,6 @@
 from .base_node import BaseNode
 
-from ..models.graph_state import GraphState
+from ..modules.graph_state import GraphState
 
 
 class GetExampleNode(BaseNode):
