+++ src/nodes/perform_rag.py
@@ -1,6 +1,6 @@
 from .base_node import BaseNode
 
-from ..models.graph_state import GraphState
+from ..modules.graph_state import GraphState
 from ..utils.data_utils import format_docs_with_metadata, format_dataframe
